reduce size of Array!T from 32 to 16 bytes (#23404)
- allocates a maximum of uint.max entries
- overlaps single entry with pointer to array

+++ compiler/src/dmd/dinterpret.d
@@ -6903,7 +6903,7 @@ private Expression interpret_aaDel(UnionExp* pue, InterState* istate, Expression
     AssocArrayLiteralExp aae = agg.isAssocArrayLiteralExp();
     Expressions* keysx = aae.keys;
     Expressions* valuesx = aae.values;
-    size_t removed = 0;
+    uint removed = 0;
     foreach (j, evalue; *valuesx)
     {
         Expression ekey = (*keysx)[j];
